test(rust): remove redundant coverage
