+++ crates/oore/tests/cli_integration.rs
